+++ src/main/java/org/folio/rest/impl/RequestsAPI.java
@@ -1,0 +1,162 @@
+package org.folio.rest.impl;
+
+import io.vertx.core.AsyncResult;
+import io.vertx.core.Context;
+import io.vertx.core.Handler;
+import org.folio.rest.jaxrs.model.Request;
+import org.folio.rest.jaxrs.resource.LoanPolicyStorageResource;
+import org.folio.rest.jaxrs.resource.RequestStorageResource;
+import org.folio.rest.persist.PostgresClient;
+import org.folio.rest.tools.utils.OutStream;
+import org.folio.rest.tools.utils.TenantTool;
+
+import javax.ws.rs.core.Response;
+import java.util.Map;
+import java.util.UUID;
+
+import static org.folio.rest.impl.Headers.TENANT_HEADER;
+
+public class RequestsAPI implements RequestStorageResource {
+
+  private final String REQUEST_TABLE = "request";
+
+  @Override
+  public void deleteRequestStorageRequests(
+    String lang,
+    Map<String, String> okapiHeaders,
+    Handler<AsyncResult<Response>> asyncResultHandler,
+    Context vertxContext) throws Exception {
+
+    String tenantId = okapiHeaders.get(TENANT_HEADER);
+
+    vertxContext.runOnContext(v -> {
+      try {
+        PostgresClient postgresClient = PostgresClient.getInstance(
+          vertxContext.owner(), TenantTool.calculateTenantId(tenantId));
+
+        postgresClient.mutate(String.format("TRUNCATE TABLE %s_%s.%s",
+          tenantId, "circulation_storage", REQUEST_TABLE),
+          reply -> {
+            asyncResultHandler.handle(io.vertx.core.Future.succeededFuture(
+              DeleteRequestStorageRequestsResponse.withNoContent()));
+          });
+      }
+      catch(Exception e) {
+        asyncResultHandler.handle(io.vertx.core.Future.succeededFuture(
+          DeleteRequestStorageRequestsResponse
+            .withPlainInternalServerError(e.getMessage())));
+      }
+    });
+  }
+
+  @Override
+  public void getRequestStorageRequests(
+    int offset,
+    int limit,
+    String lang,
+    Map<String, String> okapiHeaders,
+    Handler<AsyncResult<Response>> asyncResultHandler,
+    Context vertxContext) throws Exception {
+
+    asyncResultHandler.handle(io.vertx.core.Future.succeededFuture(
+      GetRequestStorageRequestsResponse.withNotImplemented()));
+  }
+
+  @Override
+  public void postRequestStorageRequests(
+    String lang,
+    Request entity,
+    Map<String, String> okapiHeaders,
+    Handler<AsyncResult<Response>> asyncResultHandler,
+    Context vertxContext) throws Exception {
+
+    String tenantId = okapiHeaders.get(TENANT_HEADER);
+
+    try {
+      PostgresClient postgresClient =
+        PostgresClient.getInstance(
+          vertxContext.owner(), TenantTool.calculateTenantId(tenantId));
+
+      vertxContext.runOnContext(v -> {
+        try {
+          if(entity.getId() == null) {
+            entity.setId(UUID.randomUUID().toString());
+          }
+
+          postgresClient.save(REQUEST_TABLE, entity.getId(), entity,
+            reply -> {
+              try {
+                if(reply.succeeded()) {
+                  OutStream stream = new OutStream();
+                  stream.setData(entity);
+
+                  asyncResultHandler.handle(
+                    io.vertx.core.Future.succeededFuture(
+                      PostRequestStorageRequestsResponse
+                        .withJsonCreated(reply.result(), stream)));
+                }
+                else {
+                  asyncResultHandler.handle(
+                    io.vertx.core.Future.succeededFuture(
+                      PostRequestStorageRequestsResponse
+                        .withPlainInternalServerError(reply.cause().toString())));
+                }
+              } catch (Exception e) {
+                e.printStackTrace();
+                asyncResultHandler.handle(
+                  io.vertx.core.Future.succeededFuture(
+                    PostRequestStorageRequestsResponse
+                      .withPlainInternalServerError(e.getMessage())));
+              }
+            });
+        } catch (Exception e) {
+          e.printStackTrace();
+          asyncResultHandler.handle(io.vertx.core.Future.succeededFuture(
+            PostRequestStorageRequestsResponse
+              .withPlainInternalServerError(e.getMessage())));
+        }
+      });
+    } catch (Exception e) {
+      e.printStackTrace();
+      asyncResultHandler.handle(io.vertx.core.Future.succeededFuture(
+        LoanPolicyStorageResource.PostLoanPolicyStorageLoanPoliciesResponse
+          .withPlainInternalServerError(e.getMessage())));
+    }
+  }
+
+  @Override
+  public void getRequestStorageRequestsByRequestId(
+    String requestId,
+    String lang,
+    Map<String, String> okapiHeaders,
+    Handler<AsyncResult<Response>> asyncResultHandler,
+    Context vertxContext) throws Exception {
+
+    asyncResultHandler.handle(io.vertx.core.Future.succeededFuture(
+      GetRequestStorageRequestsByRequestIdResponse.withNotImplemented()));
+  }
+
+  @Override
+  public void deleteRequestStorageRequestsByRequestId(
+    String requestId,
+    String lang,
+    Map<String, String> okapiHeaders,
+    Handler<AsyncResult<Response>> asyncResultHandler,
+    Context vertxContext) throws Exception {
+
+    asyncResultHandler.handle(io.vertx.core.Future.succeededFuture(
+      DeleteRequestStorageRequestsByRequestIdResponse.withNotImplemented()));
+  }
+
+  @Override
+  public void putRequestStorageRequestsByRequestId(
+    String requestId,
+    String lang, Request entity,
+    Map<String, String> okapiHeaders,
+    Handler<AsyncResult<Response>> asyncResultHandler,
+    Context vertxContext) throws Exception {
+
+    asyncResultHandler.handle(io.vertx.core.Future.succeededFuture(
+      PutRequestStorageRequestsByRequestIdResponse.withNotImplemented()));
+  }
+}